Updated README.md
Fixed the layout oof the text so that it would all be in the same view when rendered in GitHub. Also moved the image up to the top so that its easy to see for those coming to look at the program instead of seeing it after the setup is completed. (This is why devs don't write documentation just read the code.. its the same thing but better)